+++ lib/app/installed/installed_snaps_page.dart
@@ -22,9 +22,10 @@ import 'package:software/app/common/loading_banner_grid.dart';
 import 'package:software/app/common/snap/snap_grid.dart';
 import 'package:software/app/common/snap/snap_utils.dart';
 import 'package:software/app/common/updates_splash_screen.dart';
-import 'package:software/app/common/error_page.dart';
+import 'package:software/app/common/snap/no_snaps_installed_page.dart';
 import 'package:software/app/installed/installed_model.dart';
 import 'package:software/app/updates/no_updates_page.dart';
+import 'package:software/l10n/l10n.dart';
 import 'package:yaru_icons/yaru_icons.dart';
 
 class InstalledSnapsPage extends StatefulWidget {
@@ -65,8 +66,8 @@ class _InstalledSnapsPageState extends State<InstalledSnapsPage> {
       if (model.localSnaps.isEmpty && !model.isLoadingSnapsCompleted) {
         return const LoadingBannerGrid();
       } else if (model.localSnaps.isEmpty && model.isLoadingSnapsCompleted) {
-        return const ErrorPage(
-          message: 'Nothing to see here',
+        return NoSnapsInstalledPage(
+          message: context.l10n.noSnapsInstalled,
           icon: YaruIcons.no_package_snap,
         );
       } else {
