+++ lib/l10n/app_en.arb
@@ -143,6 +143,7 @@
   "system": "System",
   "light": "Light",
   "dark": "Dark",
+  "noSnapsInstalled" : "There are no Snap applications installed on your system",
   "appstreamSearchGreylist" : "app;application;package;program;programme;suite;tool",
   "@appstreamSearchGreylist" : {
     "description": "List of 'grey-listed' words separated with ';'. Do not translate this list directly. Instead, provide a list of words in your language that people are likely to include in a search but that should normally be ignored in the search."
